code for error handling

+++ Application/Activities/DeleteActivity.cs
@@ -1,4 +1,5 @@
-﻿using MediatR;
+﻿using Application.Core;
+using MediatR;
 using Persistance;
 using System;
 using System.Collections.Generic;
@@ -11,12 +12,12 @@ namespace Application.Activities
 {
     public class DeleteActivity
     {
-        public class Command: IRequest<Unit>
+        public class Command: IRequest<Result<Unit>>
         {
             public Guid id { get; set; }
         }
 
-        public class Handler:IRequestHandler<Command, Unit>
+        public class Handler:IRequestHandler<Command, Result<Unit>>
         {
             private readonly DataContext _context;
             public Handler(DataContext context)
@@ -24,12 +25,14 @@ namespace Application.Activities
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.id);
+                if (activity == null) return null;
                 _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
-                return Unit.Value;
+                var result = await _context.SaveChangesAsync() > 0;
+                if (!result) return Result<Unit>.Failure("Failed to delete activity");
+                return Result<Unit>.Success(Unit.Value);
             }
         }
     }
